feat: add arm-unknown-linux-gnueabihf cross-compilation support
Adds a complete build pipeline for producing a rockboxd binary targeting
ARM hard-float Linux (e.g. Raspberry Pi Zero) from a macOS host.

Changes:
- Cross.toml + Dockerfile.arm-unknown-linux-gnueabihf: cross-rs Docker
  image with arm-linux-gnueabihf-gcc, ARM dev libs, and a modern protoc
- scripts/build-armhf.sh: end-to-end build script — Docker firmware Make
  (configure target 208), cross Rust (fts5 + cpal-sink, no typesense),
  Zig final link with -Dtarget=arm-linux-gnueabihf -Dcpu=arm1176jzf_s
- tools/configure: new armhfhostcc() function and target 208 (armhfhost)
  reusing the headless target files with the ARM cross-compiler
- firmware/export/config/armhfhost.h: platform config for target 208
- firmware/export/config.h + audiohw.h + filesystem-app.c: add ARMHFHOST
  guards alongside HEADLESSHOST so the ARM build uses the same hosted
  paths (sdl_codec.h, rbhome pointer, system-target.h)
- zig/build.zig: -Dfw-dir, -Drust-triple, -Dsyslibs-dir options for
  cross-compiled firmware dir, Rust triple lib path, and ARM sysroot libs
- crates/sys: cast c_long/c_ulong FFI params through their platform type
  so the wrappers compile on 32-bit ARM (where c_long = i32, not i64)
- crates/cli/src/lib.rs: use libc::rlim_t instead of u64 for rlimit fields

+++ crates/sys/src/metadata.rs
@@ -1,4 +1,4 @@
-use std::ffi::{c_uchar, CStr, CString};
+use std::ffi::{c_long, c_uchar, c_ulong, CStr, CString};
 use std::sync::Mutex;
 
 use crate::{types::mp3_entry::Mp3Entry, ProgressFunc};
@@ -58,12 +58,12 @@ pub fn create_xing_header(
     unsafe {
         crate::create_xing_header(
             fd,
-            startpos,
-            filesize,
+            startpos as c_long,
+            filesize as c_long,
             buf,
-            num_frames,
-            rec_time,
-            header_template,
+            num_frames as c_ulong,
+            rec_time as c_ulong,
+            header_template as c_ulong,
             progressfunc,
             generate_toc,
             tempbuf,
